requirements.txt
feedparser==6.0.11
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==1.3.1
tweepy==4.14.0
transformers==4.44.0
huggingface-hub==0.24.5
tokenizers==0.19.1
torch==2.4.0
schedule==1.2.2
tqdm==4.66.5
python-dateutil==2.9.0.post0
asgiref==3.8.1
pytz==2024.1
certifi==2024.7.4
charset-normalizer==3.3.2
six==1.16.0
nltk~=3.9.1